provisioning/pyproject.toml
[project]
name = "beta9-provisioning"
version = "0.1.0"
description = "Beta9 cloud provisioning API with CDKTF"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "cdktf>=0.19.0",
    "cdktf-cdktf-provider-aws>=19.0.0",
    "cdktf-cdktf-provider-google>=11.0.0",
    "boto3>=1.34.0",
    "google-cloud-storage>=2.14.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
